+++ src/scripts/api/index.js
@@ -1,0 +1,2 @@
+export * from "./apicall";
+export * from "./config";
