--- src/scripts/footer.js
@@ -1,4 +1,0 @@
-// const input = document.getElementById("email");
-// input.addEventListener("submit", (e)=>{
-//     e.preventDefault()
-// });
