+++ src/scripts/helper/index.js
@@ -1,0 +1,1 @@
+export * from "./restaurant";
