+++ src/scripts/pages/index.js
@@ -1,0 +1,5 @@
+export * from "./favourite";
+export * from "./home";
+export * from "./not-found";
+export * from "./resto";
+export * from "./resto detail";
