+++ src/scripts/pages/resto/slider-controller.css
