+++ src/scripts/routing/index.js
@@ -1,0 +1,2 @@
+export * from "./router";
+export * from "./urlparser";
