+++ src/scripts/routing/urlparser.js
@@ -1,4 +1,4 @@
-const UrlParser = {
+export const UrlParser = {
   parseActiveUrlWithCombiner() {
     const url = window.location.hash.slice(1).toLowerCase();
     const splitedUrl = this._urlSplitter(url);
@@ -27,5 +27,3 @@ const UrlParser = {
     );
   },
 };
-
-export default UrlParser;
