--- src/scripts/topnav-links.js
@@ -1,5 +1,0 @@
-const links = document.querySelectorAll('#topnav .links');
-links.forEach((list, i) => {
-  const link = list;
-  link.style.transitionDelay = `.${i}s`;
-});
